refactoring the existing implementation
