--- src/Core/OrganizationFeatures/OrganizationSubscriptionUpdate/Interface/IAdjustSeatsCommand.cs
@@ -1,9 +1,0 @@
-﻿using Bit.Core.Entities;
-
-namespace Bit.Core.OrganizationFeatures.OrganizationSubscriptionUpdate.Interface;
-
-public interface IAdjustSeatsCommand
-{
-    Task<string> AdjustSeatsAsync(Organization organization, int seatAdjustment,
-        DateTime? prorationDate = null, IEnumerable<string> ownerEmails = null);
-}
