--- src/Core/OrganizationFeatures/OrganizationSubscriptionUpdate/Interface/IAdjustServiceAccountsCommand.cs
@@ -1,9 +1,0 @@
-﻿using Bit.Core.Entities;
-
-namespace Bit.Core.OrganizationFeatures.OrganizationSubscriptionUpdate.Interface;
-
-public interface IAdjustServiceAccountsCommand
-{
-    Task<string> AdjustServiceAccountsAsync(Organization organization, int serviceAccountAdjustment,
-        IEnumerable<string> ownerEmails = null, DateTime? prorationDate = null);
-}
