--- src/Core/OrganizationFeatures/OrganizationSubscriptionUpdate/Interface/IUpdateSeatsAutoscalingCommand.cs
@@ -1,8 +1,0 @@
-﻿using Bit.Core.Entities;
-
-namespace Bit.Core.OrganizationFeatures.OrganizationSubscriptionUpdate.Interface;
-
-public interface IUpdateSeatsAutoscalingCommand
-{
-    Task UpdateSeatsAutoscalingAsync(Organization organization, int? maxAutoscaleSeats);
-}
