--- src/Core/OrganizationFeatures/OrganizationSubscriptionUpdate/Interface/IUpdateServiceAccountAutoscalingCommand.cs
@@ -1,8 +1,0 @@
-﻿using Bit.Core.Entities;
-
-namespace Bit.Core.OrganizationFeatures.OrganizationSubscriptionUpdate.Interface;
-
-public interface IUpdateServiceAccountAutoscalingCommand
-{
-    Task UpdateServiceAccountAutoscalingAsync(Organization organization, int? maxAutoscaleServiceAccounts);
-}
